pyproject.toml
[tool.poetry]
name = "speed_tracker"
version = "0.1.0"
description = "Check your network speed and notify you via email if the network upload and download speeds are below the predefined speeds in the 'speedTracker.ini' file."
authors = ["Govine Jarrett <devgovine@gmail.com>"]
license = "None"

[tool.poetry.dependencies]
python = "^3.10"
speedtest-cli = "^2.1.3"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
